MathJax: Add condition for Math expression parsing
We previously thought (and Khanmigo is prompted) that expression
should use `\(...\)` - but in real life we also have seen `\[...\]`
for the wrapping of MathJax.

This is a simple fix for that, as well as adding a test case for it
in our task.json ensure it works.

Test Plan:

Run `yarn start` to ensure that expression display, as well as highlighting
works with the other type of expressions.